+++ N11Case/Helper/CollectionViewLayout/CollectionViewLayoutProvider.swift
@@ -131,33 +131,3 @@ final class ItemListCollectionViewLayoutProvider: CollectionViewLayoutProvider {
 }
 
 
-class SectionPagerView: UICollectionReusableView {
-    static let reuseIdentifier = "SectionPagerView"
-
-    private let pageControl: UIPageControl = {
-        let pageControl = UIPageControl()
-        pageControl.translatesAutoresizingMaskIntoConstraints = false
-        pageControl.currentPageIndicatorTintColor = .black
-        pageControl.pageIndicatorTintColor = .lightGray
-        return pageControl
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addSubview(pageControl)
-
-        NSLayoutConstraint.activate([
-            pageControl.centerXAnchor.constraint(equalTo: centerXAnchor),
-            pageControl.centerYAnchor.constraint(equalTo: centerYAnchor),
-        ])
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func configure(with numberOfPages: Int, currentPage: Int) {
-        pageControl.numberOfPages = numberOfPages
-        pageControl.currentPage = currentPage
-    }
-}
